+++ zcash_client_backend/src/data_api/testing/transparent.rs
@@ -459,7 +459,7 @@ where
 
     let external_taddrs = st
         .wallet()
-        .get_transparent_receivers(account_uuid, false)
+        .get_transparent_receivers(account_uuid, false, true)
         .unwrap();
     assert_eq!(
         u32::try_from(external_taddrs.len()).unwrap(),
@@ -467,7 +467,7 @@ where
     );
     let internal_taddrs = st
         .wallet()
-        .get_transparent_receivers(account_uuid, true)
+        .get_transparent_receivers(account_uuid, true, false)
         .unwrap();
     assert_eq!(
         u32::try_from(internal_taddrs.len()).unwrap(),
@@ -495,7 +495,7 @@ where
     // updates the gap limit by the index of the default Sapling receiver
     let external_taddrs = st
         .wallet()
-        .get_transparent_receivers(account_uuid, false)
+        .get_transparent_receivers(account_uuid, false, true)
         .unwrap();
     assert_eq!(
         u32::try_from(external_taddrs.len()).unwrap(),
@@ -529,7 +529,7 @@ where
     // never be inspected on wallet recovery.
     let external_taddrs = st
         .wallet()
-        .get_transparent_receivers(account_uuid, false)
+        .get_transparent_receivers(account_uuid, false, true)
         .unwrap();
     assert_eq!(
         u32::try_from(external_taddrs.len()).unwrap(),
@@ -547,7 +547,7 @@ where
     // Now that the transaction has been mined, the gap limit should have increased.
     let external_taddrs = st
         .wallet()
-        .get_transparent_receivers(account_uuid, false)
+        .get_transparent_receivers(account_uuid, false, true)
         .unwrap();
     assert_eq!(
         u32::try_from(external_taddrs.len()).unwrap(),